+++ src/login/app_config/login-rootSaga.js
@@ -1,6 +1,7 @@
 import { takeLatest } from "redux-saga/effects";
 import * as init_actions from "../app_init/init_actions";
-import { requestConfig, getResetPasswordData, postEmail } from "../app_init/init_sagas";
+import { requestConfig } from "../app_init/init_sagas";
+import { getResetPasswordData, postEmail } from "../redux/sagas/resetpassword/resetPasswordSaga";
 import * as resetPasswordActions from "../redux/actions/resetPasswordActions";
 
 function* rootSaga() {
